+++ NEWS.md
@@ -14,6 +14,8 @@
 
 * Fix simplification of inverse units; #378
 
+* Replace call to `Rf_error()` with `Rcpp::stop()`; RcppCore/Rcpp#1247
+
 # version 0.8-5
 
 * avoid -Wformat-security warning on CRAN