+++ NEWS.md
@@ -16,6 +16,8 @@
 
 * Replace call to `Rf_error()` with `Rcpp::stop()`; RcppCore/Rcpp#1247
 
+* Fix UBs in the C++ glue code; #380
+
 # version 0.8-5
 
 * avoid -Wformat-security warning on CRAN